Split Animation into SequenceAnimation and ClipAnimation
