+++ Tests/AnimationTest.cs
@@ -35,6 +35,7 @@ namespace Tests
             {
                 Assert.That(animGroupPaths.Count(), Is.Zero);
                 Assert.That(animations, Has.Count.EqualTo(3));
+                Assert.That(animations, Is.All.InstanceOf<SequenceAnimation>());
 
                 Assert.That(animations[0].Name, Is.EqualTo("ref_pose"));
                 Assert.That(animations[0].Fps, Is.EqualTo(30));
