+++ ValveResourceFormat/IO/ModelExtract.Anim.cs
@@ -13,7 +13,7 @@ partial class ModelExtract
     /// <summary>
     /// Gets the list of animations to be extracted with their output file names.
     /// </summary>
-    public List<(Animation Anim, string FileName)> AnimationsToExtract { get; } = [];
+    public List<(SequenceAnimation Anim, string FileName)> AnimationsToExtract { get; } = [];
 
     private void EnqueueAnimations()
     {
